server/requirements.txt
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
databases==0.5.5
et-xmlfile==1.1.0
fake-useragent==0.1.11
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
lxml==4.8.0
openpyxl==3.0.9
pathlib2==2.3.7.post1
pydantic==1.9.0
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.35
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
watchgod==0.7
websockets==10.2
